requirements.txt
asgiref==3.5.2
astroid==2.12.10
autopep8==1.7.0
boto3==1.24.84
botocore==1.27.84
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-datetime==0.1
django-environ==0.9.0
django-filter==22.1
django-model-utils==4.2.0
django-simple-pwa==2.2.1
django-storages==1.13.1
django-tinymce==3.5.0
docopt==0.6.2
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mercadopago==2.2.0
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.1
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
types-cryptography==3.3.23
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
yarg==0.1.9
zope.interface==5.4.0
